--- assets/RecordButton.css
@@ -1,99 +1,0 @@
-body {
-	font-family: -apple-system, BlinkMacSystemFont, "Helvetica Neue", "メイリオ", Meiryo,
-		"ヒラギノ角ゴ ProN W3", "Hiragino Kaku Gothic ProN", Arial, sans-serif;
-	-webkit-text-size-adjust: none;
-	line-height: 1.5;
-}
-/*** 記録 ***/
-.drecord-record-button-container {
-	position: absolute;
-	bottom: 0;
-	right: 0;
-	display: flex;
-	pointer-events: none;
-}
-.drecord-record-button-list {
-	display: flex;
-	margin: 3px 0;
-	opacity: 0;
-	visibility: hidden;
-	transform: translateX(15px);
-	transition: 0.15s ease-in-out;
-	pointer-events: auto;
-}
-.drecord-record-button-list.show {
-	opacity: 1;
-	visibility: visible;
-	transform: none;
-}
-.drecord-record-button {
-	display: flex;
-	align-items: center;
-	border: 1px solid rgba(0, 0, 0, 0.2);
-	border-radius: 2px;
-	white-space: nowrap;
-	margin-right: 4px;
-	padding: 1px 4px;
-	color: #5b5b5b;
-	cursor: pointer;
-	transition: 0.15s ease-in-out;
-	background-color: #fff;
-	user-select: none;
-}
-.drecord-record-button:nth-child(2) {
-	margin-right: 2px;
-}
-.drecord-record-button > span {
-	font-size: 14px;
-	line-height: 1;
-}
-@media screen and (min-width: 960px) {
-	.drecord-record-button > span {
-		font-size: 12px;
-	}
-}
-.drecord-record-button:hover {
-	background-color: rgb(225, 227, 231);
-}
-.drecord-record-button:active {
-	background-color: rgb(179, 180, 184);
-}
-.drecord-record-svg-container {
-	padding: 4px 4px 4px 0;
-	cursor: pointer;
-	transition: transform 0.2s;
-	pointer-events: auto;
-}
-.drecord-record-svg-container:active {
-	transform: scale(0.9);
-}
-.drecord-record-svg {
-	position: relative;
-	width: 30px;
-	height: 30px;
-	z-index: 100;
-}
-@media screen and (min-width: 960px) {
-	.drecord-record-svg {
-		width: 25px;
-		height: 25px;
-	}
-}
-
-.drecord-record-svg > path {
-	transition: 0.25s ease-in-out;
-}
-.drecord-record-svg > path:nth-child(1),
-.drecord-record-svg > path:nth-child(2) {
-	opacity: 0;
-}
-.drecord-record-svg > path:nth-child(1).show,
-.drecord-record-svg > path:nth-child(2).show {
-	opacity: 1;
-}
-
-.drecord-record-button > svg {
-	width: 17px;
-	height: 17px;
-	margin-right: 1px;
-}
