--- assets/UploadToggleButton.css
@@ -1,108 +1,0 @@
-/*** 再生画面 ***/
-#drecord-container {
-	display: flex;
-	width: auto !important;
-}
-
-/* アイコン */
-#upload-icon-container {
-	position: relative;
-	width: 44px;
-	display: flex;
-	justify-content: center;
-	align-items: center;
-	cursor: pointer;
-}
-#upload-icon-container:hover > #upload-icon {
-	opacity: 1;
-}
-#upload-icon {
-	opacity: 0.6;
-}
-.not-upload-icon {
-	opacity: 0.3 !important;
-}
-.drop-shadow {
-	filter: drop-shadow(0 0 9px #fff);
-}
-
-#drecord-popup {
-	right: -8px;
-	pointer-events: none;
-	position: absolute;
-	top: 0;
-	transform: translateY(10px);
-	opacity: 0;
-	transition:
-		transform 0.2s cubic-bezier(0, 0, 0.2, 1),
-		opacity 0.2s cubic-bezier(0, 0, 0.2, 1);
-	z-index: 90;
-	box-sizing: border-box;
-}
-#upload-icon-container:hover > #drecord-popup {
-	transform: none;
-	opacity: 1;
-}
-#drecord-popupIn {
-	position: relative;
-	transform: translateY(-100%);
-	padding: 0;
-	background-color: rgba(0, 0, 0, 0.8);
-	border-radius: 4px;
-	bottom: 22px;
-	box-shadow: 0 0 6px 4px rgba(64, 64, 64, 0.35);
-}
-#drecord-popupIn::after {
-	content: "";
-	position: absolute;
-	z-index: 90;
-	bottom: -7px;
-	margin-left: -6px;
-	display: block;
-	width: 0;
-	height: 0;
-	border-style: solid;
-	border-width: 7px 6px 0 6px;
-	border-color: rgba(0, 0, 0, 0.8) transparent;
-	right: 24px;
-}
-#drecord-popupIn > span {
-	display: inline-block;
-	color: rgb(160, 160, 159);
-	font-size: 12px;
-	font-weight: 700;
-	white-space: nowrap;
-	padding: 0.5em 0.7em;
-}
-
-/* アニメタイトル */
-#upload-anime-title {
-	display: block;
-	transition: 0.5s;
-	max-width: 0px;
-	opacity: 0;
-	visibility: hidden;
-}
-#upload-anime-title.show {
-	max-width: 500px;
-	opacity: 1;
-	visibility: visible;
-}
-
-#upload-anime-title > span {
-	display: flex !important;
-	height: 100%;
-	color: #ffffff;
-	font-weight: 700;
-	align-items: center;
-	opacity: 0;
-	text-overflow: ellipsis;
-	overflow: hidden;
-	white-space: nowrap;
-	transition: 0.4s;
-	visibility: hidden;
-}
-#upload-anime-title > span.show {
-	opacity: 0.6;
-	visibility: visible;
-}
