+++ entrypoints/content/components/common/UploadToggleButton.tsx
@@ -1,11 +1,10 @@
 /* 再生ページのアップロードボタン */
-
 import AnimeTitle from "./AnimeTitle";
-import { AbemaUploadToggleButton } from "./abema/AbemaUploadToggleButton";
-import { DanimeUploadToggleButton } from "./danime/DanimeUploadToggleButton";
-import { loading, setLoading } from "../anime-data-scraper";
-import { cleanupIntervalOrEvent, createIntervalOrEvent } from "../record-watch-episode";
-import { getNotRecordWork, settingData } from "../storage";
+import { AbemaUploadToggleButton } from "../abema/AbemaUploadToggleButton";
+import { DanimeUploadToggleButton } from "../danime/DanimeUploadToggleButton";
+import { loading, setLoading } from "../../core/anime-data-scraper";
+import { cleanupIntervalOrEvent, createIntervalOrEvent } from "../../core/record-watch-episode";
+import { getNotRecordWork, settingData } from "../../utils/storage";
 
 const uploadSVGIcon =
 	"data:image/svg+xml,%3Csvg%20xmlns%3D%22http%3A%2F%2Fwww.w3.org%2F2000%2Fsvg%22%20xml%3Aspace%3D%22preserve%22%20viewBox%3D%220%200%2024%2024%22%3E%3Cpath%20fill%3D%22%23FFF%22%20d%3D%22M1%2015.5c0-2.3%201.2-4.4%203.1-5.5.5-4%203.8-7%207.9-7%204.1%200%207.4%203%207.9%207%201.8%201.1%203.1%203.2%203.1%205.5%200%203.4-2.6%206.2-6%206.5H7c-3.4-.3-6-3.1-6-6.5zM16.8%2020c2.3-.2%204.2-2.1%204.2-4.5%200-1.6-.8-3-2.1-3.8l-.8-.5-.1-1C17.6%207.3%2015%205%2012%205s-5.6%202.3-6%205.2l-.1.9-.8.5c-1.3.9-2.1%202.3-2.1%203.9%200%202.4%201.8%204.3%204.2%204.5h9.6z%22%2F%3E%3Cpath%20fill%3D%22%23FFF%22%20d%3D%22M13%2013v4h-2v-4H8l4-5%204%205z%22%2F%3E%3C%2Fsvg%3E";
