+++ entrypoints/popup/index.html
@@ -7,8 +7,8 @@
 		<link rel="stylesheet" href="~/assets/tailwind.css" />
 	</head>
 
-	<body class="w-[350px]">
-		<div id="root" class="[&_h2]:mb-3"></div>
+	<body>
+		<div id="root"></div>
 		<script type="module" src="./main.tsx"></script>
 	</body>
 </html>
