+++ entrypoints/popup/main.tsx
@@ -1,4 +1,3 @@
-import { render } from "solid-js/web";
 import App from "./App";
 
 render(() => <App />, document.getElementById("root")!);
