+++ entrypoints/popup/type.ts
@@ -1,12 +1,10 @@
 export type Settings = {
-	Token?: string;
-	sendTiming?: string;
-	nextEpisodeLine?: boolean;
-	recordButton?: boolean;
-	animeTitle?: boolean;
-	autoChangeStatus?: boolean;
-	applyWebsite?: {
-		danime?: boolean;
-		abema?: boolean;
+	sendTiming: string;
+	nextEpisodeLine: boolean;
+	recordButton: boolean;
+	animeTitle: boolean;
+	autoChangeStatus: boolean;
+	applyWebsite: {
+		[key: string]: boolean;
 	};
 };
